+++ packages/adapter-launchdarkly/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @flags-sdk/launchdarkly
 
+## 0.3.0
+
+### Minor Changes
+
+- 917ef42: change API from ldAdapter() to ldAdapter.variation()
+
+### Patch Changes
+
+- b375e4e: add metadata to package.json
+- 917ef42: expose ldClient on default ldAdapter
+
+## 0.2.1
+
+### Patch Changes
+
+- fbc886a: expose ldAdapter.ldClient
+- fbc886a: expose as ldAdapter
+
+## 0.2.0
+
+### Minor Changes
+
+- 2fcc446: Add LaunchDarkly adapter
+
 ## 0.1.0
 
 ### Minor Changes